[NFC] Move non-experimental tensor descritor tests to their own file
git-pr-chain: nfc_move_non_experimental_tensor_descrit_1b70
